requirements.txt
fastapi
uvicorn[standard]
python-jobspy
pydantic
pydantic-settings
python-multipart
psutil
python-dotenv
requests
beautifulsoup4
